feat: add key package batch builder for virtual clients
The newly added `VcKeyPackageBatchBuilder` allows to build heterogeneous
batches of key packages. Build the batch on `finalize`. Dropping the
builder without calling `finalize` burns no generation.
